// fsiem_cloud/src/portal/ui/src/app/models/application-route.ts
export enum ApplicationRoute {
    NULL = '',
    PRODUCT_INFORMATION = 'product-information',
    LOGIN = 'login',
    LOGOUT = 'logout',
    ENTITLEMENTS = 'entitlements',
    DEPLOYMENT_DETAILS = 'deployment-details',
    ACCOUNT_SELECTED = 'accountselected',
    ACCOUNT_SELECTION = 'account-selection',
    SPLASH = 'splash'
}